+++ flutter_feature_biometric_platform_interface/lib/src/enum/feature_authenticator_status.dart
@@ -1,4 +1,4 @@
-enum AuthenticatorStatus {
+enum FeatureAuthenticatorStatus {
   success,
   noHardwareAvailable,
   unavailable,