+++ flutter_feature_biometric_platform_interface/lib/src/state/success_authenticate_encrypt_state.dart
@@ -10,4 +10,8 @@ class SuccessAuthenticateEncryptAndroid extends SuccessAuthenticateEncryptState 
   });
 }
 
-class SuccessAuthenticateEncryptIOS extends SuccessAuthenticateEncryptState {}
+class SuccessAuthenticateEncryptIOS extends SuccessAuthenticateEncryptState {
+  String encodedDomainState;
+
+  SuccessAuthenticateEncryptIOS({required this.encodedDomainState});
+}