+++ flutter_feature_biometric_platform_interface/CHANGELOG.md
@@ -1,5 +1,13 @@
 ## 0.1.0
 
+* add parameter `encodedKey` inside `isBiometricChanged`
+* change `key` to `alias` inside `authenticateSecureEncrypt`
+* change `key` to `alias` inside `authenticateSecureDecrypt`
+* change `encodedIVKey` to `encodedKey` inside `authenticateSecureDecrypt`
+* change `AuthenticatorStatus` to `FeatureAuthenticatorStatus`
+
+## 0.1.0
+
 * Update Docs
 * Update authentication biometric & authentication device credential using same function